Move backend code to backend directory
Move all relevant go code for GitHub API querying and chart data aggregation
into `backend` directory. Only update `backend/cmd/leaderfetcher/main.go`
imports.

Leave Makefile and cloudbuild.yaml in root directory.
